split Puppeteer class
