+++ lib/puppeteer/version.rb
@@ -1,3 +1,3 @@
-class Puppeteer
+module Puppeteer
   VERSION = '0.31.0'
 end
